fix(analytics): clear detekt and Sonar findings from glossary wiring
Extract share card rendering and telemetry helpers so LargeClass/complexity stay under threshold; dedupe OPML error string for Sonar.
