+++ config/detekt/baseline.xml
@@ -99,7 +99,7 @@
     <ID>CyclomaticComplexMethod:RssPodcastRepository.kt$RssPodcastRepository$private suspend fun migrateLinkedState( podcastIndexPodcast: PodcastEntity, rssPodcast: PodcastEntity, rssEpisodes: List&lt;RssEpisodeEntity&gt;, )</ID>
     <ID>CyclomaticComplexMethod:SearchOnboardingScreen.kt$@Composable internal fun OnboardingSearchScreen( query: String, results: List&lt;Podcast&gt;, isSearching: Boolean, subscribedIds: Set&lt;String&gt;, selectedPodcasts: Map&lt;String, Podcast&gt;, onQueryChange: (String) -&gt; Unit, onSubscribe: (Podcast) -&gt; Unit, onBack: () -&gt; Unit, onDone: () -&gt; Unit, popularPodcasts: List&lt;Podcast&gt;, isPopularLoading: Boolean, selectedSearchGenre: String?, onGenreSelect: (String?) -&gt; Unit, isAiSideTrip: Boolean = false, )</ID>
     <ID>CyclomaticComplexMethod:SessionAggregator.kt$PlayerSessionAggregator$fun endSession()</ID>
-    <ID>CyclomaticComplexMethod:ShareManager.kt$ShareManager$private fun drawShareCardShapes( canvas: android.graphics.Canvas, width: Float, height: Float, isStory: Boolean )</ID>
+    <ID>CyclomaticComplexMethod:ShareCardRenderer.kt$ShareCardRenderer$private fun drawShareCardShapes( canvas: android.graphics.Canvas, width: Float, height: Float, isStory: Boolean )</ID>
     <ID>CyclomaticComplexMethod:SmartDownloadManager.kt$SmartDownloadManager$private suspend fun triggerDownloads( combinedEpisodes: List&lt;Episode&gt;, existingDownloads: List&lt;DownloadedEpisodeEntity&gt;, subs: List&lt;PodcastEntity&gt;, maxCount: Int, storageBudgetMb: Long, startingDownloadedBytes: Long, startingCount: Int )</ID>
     <ID>CyclomaticComplexMethod:SmartDownloadsDashboard.kt$@Composable fun SmartDownloadsDashboardCard( userPrefs: cx.aswin.boxlore.core.prefs.UserPreferencesRepository, downloads: List&lt;cx.aswin.boxlore.core.database.DownloadedEpisodeEntity&gt;, onSettingsClick: () -&gt; Unit, onSyncNow: () -&gt; Unit, isSyncing: Boolean = false )</ID>
     <ID>CyclomaticComplexMethod:SmartDownloadsSettingsScreen.kt$@OptIn(ExperimentalMaterial3Api::class) @Composable fun SmartDownloadsSettingsScreen( userPrefs: UserPreferencesRepository, onBack: () -&gt; Unit )</ID>
@@ -297,9 +297,9 @@
     <ID>LongParameterList:PodcastRepository.kt$PodcastRepository$( episodeId: String, podcastId: String, title: String, description: String, podcastTitle: String, categories: String = "", author: String = "", limit: Int = 10, country: String? = null )</ID>
     <ID>LongParameterList:SearchOnboardingScreen.kt$( query: String, results: List&lt;Podcast&gt;, isSearching: Boolean, subscribedIds: Set&lt;String&gt;, selectedPodcasts: Map&lt;String, Podcast&gt;, onQueryChange: (String) -&gt; Unit, onSubscribe: (Podcast) -&gt; Unit, onBack: () -&gt; Unit, onDone: () -&gt; Unit, popularPodcasts: List&lt;Podcast&gt;, isPopularLoading: Boolean, selectedSearchGenre: String?, onGenreSelect: (String?) -&gt; Unit, isAiSideTrip: Boolean = false, )</ID>
     <ID>LongParameterList:ShareBottomSheet.kt$( id: String, type: String, // "podcast" or "episode" title: String, subtitle: String, imageUrl: String? = null, onDismissRequest: () -&gt; Unit, durationMs: Long = 0L, currentPositionMs: Long = 0L, showTimestampOption: Boolean = false, onShare: ( id: String, type: String, timestampMs: Long?, target: ShareTarget ) -&gt; Unit )</ID>
-    <ID>LongParameterList:ShareManager.kt$ShareManager$( canvas: android.graphics.Canvas, canvasWidth: Int, text: String, top: Float, width: Int, textSize: Float, maxLines: Int, color: Int, typeface: android.graphics.Typeface )</ID>
-    <ID>LongParameterList:ShareManager.kt$ShareManager$( context: Context, canvas: android.graphics.Canvas, canvasWidth: Int, brandingTop: Float, brandingLabelSize: Float, logoWidth: Int, listenNowGap: Float, listenNowSize: Float )</ID>
-    <ID>LongParameterList:ShareManager.kt$ShareManager$( context: Context, canvas: android.graphics.Canvas, canvasWidth: Int, title: String, subtitle: String, titleTop: Float, titleWidth: Int, titleSize: Float, subtitleSize: Float, blockSpacing: Float )</ID>
+    <ID>LongParameterList:ShareCardRenderer.kt$ShareCardRenderer$( canvas: android.graphics.Canvas, canvasWidth: Int, text: String, top: Float, width: Int, textSize: Float, maxLines: Int, color: Int, typeface: android.graphics.Typeface )</ID>
+    <ID>LongParameterList:ShareCardRenderer.kt$ShareCardRenderer$( context: Context, canvas: android.graphics.Canvas, canvasWidth: Int, brandingTop: Float, brandingLabelSize: Float, logoWidth: Int, listenNowGap: Float, listenNowSize: Float )</ID>
+    <ID>LongParameterList:ShareCardRenderer.kt$ShareCardRenderer$( context: Context, canvas: android.graphics.Canvas, canvasWidth: Int, title: String, subtitle: String, titleTop: Float, titleWidth: Int, titleSize: Float, subtitleSize: Float, blockSpacing: Float )</ID>
     <ID>LongParameterList:ShareManager.kt$ShareManager$( context: Context, text: String, shareUrl: String, title: String, cardTitle: String, cardSubtitle: String, imageUrl: String?, fallbackImageUrl: String? = null, target: ShareTarget )</ID>
     <ID>LongParameterList:SmartQueueRefillCoordinator.kt$SmartQueueRefillCoordinator$( private val database: BoxLoreDatabase, private val podcastRepository: PodcastRepository, private val queueRepository: QueueRepository, private val smartQueueEngine: SmartQueueEngine, private val userPreferencesRepository: UserPreferencesRepository, private val mainDispatcher: CoroutineDispatcher, private val ioDispatcher: CoroutineDispatcher, private val findPodcastIdForEpisode: suspend (String) -&gt; String?, private val queueMaxSize: Int, private val mediaIdPrefixStripper: (String) -&gt; String, )</ID>
     <ID>LongParameterList:SubscriptionTabContents.kt$( podcasts: List&lt;Podcast&gt;, allHistory: List&lt;ListeningHistoryEntity&gt;, useSmartRank: Boolean, scoreEpisodes: suspend ( List&lt;Podcast&gt;, List&lt;ListeningHistoryEntity&gt;, ) -&gt; Map&lt;String, Double&gt;, onExploreClick: () -&gt; Unit, onEpisodeClick: ((Episode, Podcast, String?) -&gt; Unit)?, onPlayEpisode: ((Episode, Podcast) -&gt; Unit)?, onPlayEpisodes: ((List&lt;Episode&gt;, Podcast) -&gt; Unit)? = null, isPlayerActive: Boolean = false )</ID>
